refactor(landing): rename NewLayout to App, kill dead code, fix broken links
- Rename pages/NewLayout.tsx → src/App.tsx (canonical homepage)
- Remove legacy App.tsx (Hero/ValueProps/Carousel/Features/Team/FAQ/CTA/Values
  layout no longer used)
- Drop /newlayout route alias
- Delete 34 dead files: 17 components + CSS, 1 proto, 2 hooks, 5 storyboard
  stages, and an empty pages/ tree
- Footer: #chronicles → URLS.blog.index, Privacy/Terms → URLS.docs.privacy
- Navbar: replace 6 broken anchors (#features/#how/#showcase/#values/#team/
  #faq) with 4 external links (Manifesto, Docs, Explorer, Discord)
- urls.ts: prune unused keys (gettingStarted, certifications, 4 logbooks,
  fromIdea, alpha, values, proxyDashboard), add docs.privacy
