--- apps/landing/src/components/CTA.module.css
@@ -1,64 +1,0 @@
-/**
- * CTA — peach panel, centered text, hex-split scroll-driven decoration.
- */
-
-.section {
-  position: relative;
-  background: var(--color-accent);
-  color: #02000e;
-  padding: 140px var(--gutter);
-  text-align: center;
-  overflow: hidden;
-}
-
-.inner {
-  position: relative;
-  z-index: 1;
-  max-width: 760px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  align-items: center;
-}
-
-.eyebrow {
-  font-family: var(--font-mono);
-  font-size: 11px;
-  letter-spacing: 0.24em;
-  color: #02000e;
-  text-transform: uppercase;
-  display: inline-flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.eyebrow::before {
-  content: '';
-  width: 6px;
-  height: 6px;
-  background: #02000e;
-}
-
-.title {
-  margin: 0 auto;
-  max-width: 18ch;
-  color: #02000e;
-}
-
-.sub {
-  font-family: var(--font-body);
-  font-size: 1.05rem;
-  line-height: 1.55;
-  color: rgba(0, 0, 0, 0.72);
-  max-width: 36rem;
-  margin: 0 auto;
-}
-
-.btns {
-  display: flex;
-  gap: 12px;
-  justify-content: center;
-  flex-wrap: wrap;
-  margin-top: 8px;
-}
