--- apps/landing/src/components/Carousel.module.css
@@ -1,219 +1,0 @@
-/**
- * Carousel (S.04) — showcase grid: stage left (image), info right
- * (counter + title + sub + pip rail + 6 thumb window).
- */
-
-.grid {
-  display: grid;
-  grid-template-columns: minmax(0, 1.7fr) minmax(280px, 340px);
-  gap: 32px;
-  align-items: center;
-  /* Fill the available height of the slide so the stage image has a
-     hard ceiling and never spills past 100vh. */
-  flex: 1;
-  min-height: 0;
-  height: 100%;
-}
-
-.stage {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  min-height: 0;
-}
-
-.imgWrap {
-  position: relative;
-  border-radius: var(--radius-lg, 14px);
-  overflow: hidden;
-  border: 1px solid var(--color-border);
-  background: var(--color-card, #0e0e0e);
-  box-shadow: 0 30px 80px rgba(0, 0, 0, 0.45);
-  aspect-ratio: 16 / 10;
-  /* Cap to the slide height minus rough chrome (section padding +
-     navbar + SectionHead). object-fit keeps the image intact. */
-  max-height: calc(100vh - var(--nav-h, 64px) - 26vh);
-  width: 100%;
-  max-width: 100%;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-  object-position: center;
-  display: block;
-  transition: opacity 0.5s var(--ease-out);
-}
-
-.arrow {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 46px;
-  height: 46px;
-  border-radius: 50%;
-  border: 1px solid var(--color-border);
-  background: color-mix(in srgb, var(--color-bg) 70%, transparent);
-  color: var(--color-text);
-  font-size: 1.4rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 3;
-  cursor: pointer;
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  transition:
-    background 0.2s,
-    border-color 0.2s,
-    color 0.2s,
-    transform 0.25s;
-}
-
-.arrow:hover {
-  background: var(--color-accent);
-  border-color: var(--color-accent);
-  color: var(--color-accent-ink);
-  transform: translateY(-50%) scale(1.06);
-}
-
-.arrowLeft {
-  left: 16px;
-}
-.arrowRight {
-  right: 16px;
-}
-
-/* ── Info column ──────────────────────────────────────── */
-
-.info {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  align-self: center;
-  min-height: 0;
-}
-
-.counter {
-  display: inline-flex;
-  align-items: baseline;
-  gap: 4px;
-  font-family: var(--font-mono);
-  font-variant-numeric: tabular-nums;
-}
-
-.counterCur {
-  font-size: 1.5rem;
-  font-weight: 500;
-  color: var(--color-text);
-  letter-spacing: -0.02em;
-}
-
-.counterTot {
-  font-size: 0.875rem;
-  color: var(--color-text-dim);
-  letter-spacing: 0.04em;
-}
-
-.counterTot::before {
-  content: '/ ';
-}
-
-.title {
-  font-family: var(--font-heading);
-  font-size: clamp(1.2rem, 0.95rem + 0.6vw, 1.6rem);
-  font-weight: 500;
-  line-height: 1.1;
-  letter-spacing: -0.024em;
-  color: var(--color-text);
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-}
-
-.sub {
-  font-size: 0.85rem;
-  line-height: 1.45;
-  color: var(--color-text-muted);
-  margin: 0;
-}
-
-.rail {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  flex-wrap: wrap;
-}
-
-.pip {
-  width: 6px;
-  height: 6px;
-  border-radius: 50%;
-  background: var(--color-border);
-  cursor: pointer;
-  border: none;
-  padding: 0;
-  transition:
-    width 0.35s var(--ease-out),
-    background 0.25s;
-}
-
-.pip:hover {
-  background: var(--color-text-muted);
-}
-
-.pipOn {
-  width: 24px;
-  border-radius: 999px;
-  background: var(--color-accent);
-}
-
-.thumbs {
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  gap: 4px;
-  margin-top: 4px;
-}
-
-.thumb {
-  border-radius: 4px;
-  overflow: hidden;
-  border: 1px solid var(--color-border-soft);
-  aspect-ratio: 16 / 10;
-  cursor: pointer;
-  background: var(--color-card, #0e0e0e);
-  opacity: 0.45;
-  transition:
-    opacity 0.25s,
-    border-color 0.25s,
-    transform 0.25s;
-  padding: 0;
-}
-
-.thumb img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: top left;
-  display: block;
-}
-
-.thumb:hover {
-  opacity: 0.75;
-}
-
-.thumbOn {
-  opacity: 1;
-  border-color: var(--color-accent);
-}
-
-@media (max-width: 720px) {
-  .grid {
-    grid-template-columns: 1fr;
-    gap: 20px;
-  }
-}
