--- apps/landing/src/components/Chronicles.module.css
@@ -1,94 +1,0 @@
-/**
- * Chronicles (S.06) — header bipartite (title left + secondary CTA right),
- * then a PLATE.D wrapping a 3-col grid of build-log entries (chrons).
- */
-
-.allLink {
-  align-self: end;
-  justify-self: end;
-}
-
-.chronsPlate {
-  width: 100%;
-}
-
-.chrons {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 0;
-  border-top: 1px solid var(--color-border-soft);
-}
-
-.chron {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-  padding: 36px 28px 40px;
-  border-right: 1px solid var(--color-border-soft);
-  border-bottom: 1px solid var(--color-border-soft);
-  transition: background 0.25s;
-  color: inherit;
-  text-decoration: none;
-}
-
-.chron:last-child {
-  border-right: 0;
-}
-
-.chron:hover {
-  background: var(--color-bg-alt);
-}
-
-.date {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  text-transform: uppercase;
-  color: var(--color-accent-ink);
-}
-
-.title {
-  font-family: var(--font-heading);
-  font-size: 1.25rem;
-  font-weight: 500;
-  letter-spacing: -0.022em;
-  line-height: 1.2;
-  color: var(--color-text);
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-}
-
-.excerpt {
-  font-size: 0.925rem;
-  line-height: 1.6;
-  color: var(--color-text-muted);
-  flex: 1;
-  margin: 0;
-}
-
-.link {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.18em;
-  text-transform: uppercase;
-  color: var(--color-text);
-  align-self: flex-start;
-}
-
-.chron:hover .link {
-  color: var(--color-accent-ink);
-}
-
-@media (max-width: 880px) {
-  .chrons {
-    grid-template-columns: 1fr;
-  }
-  .chron {
-    border-right: 0;
-  }
-}
