--- apps/landing/src/components/Comparison.module.css
@@ -1,117 +1,0 @@
-/**
- * Comparison (S.03) — peach section hosting a `cmp-dark` plate that
- * wraps the comparison table. Table is dark-on-light theming inverted:
- * white text on #0a0a0a with peach checkmarks.
- */
-
-.cmpPlate {
-  width: 100%;
-}
-
-/* ── Table (inside the dark plate) ────────────────────── */
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  font-family: var(--font-mono);
-  font-size: 0.875rem;
-  color: var(--color-text);
-}
-
-.table thead th {
-  text-align: left;
-  padding: 18px 18px 16px;
-  font-size: 11px;
-  font-weight: 500;
-  letter-spacing: 0.22em;
-  text-transform: uppercase;
-  color: var(--color-text-dim);
-  border-bottom: 1px solid var(--color-border-soft);
-}
-
-.table thead th:first-child {
-  padding-left: 24px;
-}
-
-.table thead th.headSofia {
-  color: var(--color-accent);
-  font-weight: 700;
-}
-
-.table tbody td {
-  padding: 20px 18px;
-  border-bottom: 1px solid var(--color-border-soft);
-  vertical-align: middle;
-}
-
-.table tbody tr {
-  transition: background 0.2s var(--ease-out);
-}
-
-.table tbody tr:hover {
-  background: rgba(255, 255, 255, 0.04);
-}
-
-.rowLabel {
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  text-transform: uppercase;
-  color: var(--color-text-dim);
-  width: 24%;
-  padding-left: 24px !important;
-}
-
-.tdSofia {
-  color: var(--color-accent);
-  font-weight: 500;
-}
-
-.cell {
-  display: inline-flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.check,
-.cross {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 18px;
-  height: 18px;
-  font-size: 11px;
-  font-weight: 700;
-  border-radius: 999px;
-  flex-shrink: 0;
-}
-
-.check {
-  background: var(--color-accent);
-  color: #02000e;
-  border: 1px solid var(--color-accent);
-}
-
-.cross {
-  background: transparent;
-  color: rgba(255, 255, 255, 0.4);
-  border: 1px solid rgba(255, 255, 255, 0.32);
-}
-
-/* ── Responsive ───────────────────────────────────────── */
-
-@media (max-width: 768px) {
-  .table {
-    display: block;
-    overflow-x: auto;
-    -webkit-overflow-scrolling: touch;
-    white-space: nowrap;
-  }
-  .table th,
-  .table td {
-    padding: 14px 14px;
-    font-size: 0.8125rem;
-  }
-  .rowLabel {
-    width: auto;
-  }
-}
