--- apps/landing/src/components/Features.module.css
@@ -1,350 +1,0 @@
-/**
- * Product (S.01) — three blocks on peach:
- *   1. Two-pillar overview (Explorer + Extension)
- *   2. Vision long-form
- *   3. Open & built-in-public (3-cell grid)
- *
- * Mirrors the ink-on-peach visual language used by Values + ValueProps.
- */
-
-.wide :global(.container) {
-  max-width: min(1480px, calc(100vw - 48px));
-}
-
-/* ── 1. Pillars — 2-card grid ──────────────────────────── */
-
-.pillars {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 0;
-  margin-top: 56px;
-  border-top: 1px solid rgba(0, 0, 0, 0.18);
-  border-left: 1px solid rgba(0, 0, 0, 0.18);
-}
-
-.pillar {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-  padding: 40px 36px 44px;
-  border-right: 1px solid rgba(0, 0, 0, 0.18);
-  border-bottom: 1px solid rgba(0, 0, 0, 0.18);
-  min-width: 0;
-  background: var(--color-accent);
-  transition: background 0.25s var(--ease-out);
-}
-
-.pillar:hover {
-  background: rgba(0, 0, 0, 0.04);
-}
-
-.pillarHead {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  align-items: flex-start;
-}
-
-.pillarTag {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  color: rgba(0, 0, 0, 0.55);
-}
-
-.pillarIcon {
-  width: 56px;
-  height: 56px;
-  display: grid;
-  place-items: center;
-  border: 1px solid rgba(0, 0, 0, 0.25);
-  color: #02000e;
-  transition:
-    border-color 0.25s,
-    background 0.25s;
-}
-
-.pillar:hover .pillarIcon {
-  border-color: #02000e;
-  background: rgba(0, 0, 0, 0.06);
-}
-
-.pillarIcon svg {
-  width: 26px;
-  height: 26px;
-  stroke-linecap: round;
-  stroke-linejoin: round;
-}
-
-.pillarName {
-  font-family: var(--font-heading);
-  font-size: 1.7rem;
-  font-weight: 500;
-  letter-spacing: -0.022em;
-  line-height: 1.1;
-  color: #02000e;
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-}
-
-.pillarDesc {
-  font-size: 1rem;
-  line-height: 1.55;
-  color: rgba(0, 0, 0, 0.78);
-  margin: 0;
-  max-width: 56ch;
-}
-
-.pillarBullets {
-  list-style: none;
-  margin: 4px 0 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.pillarBullets li {
-  position: relative;
-  padding-left: 22px;
-  font-size: 0.92rem;
-  line-height: 1.5;
-  color: rgba(0, 0, 0, 0.72);
-}
-
-.pillarBullets li::before {
-  content: '';
-  position: absolute;
-  left: 0;
-  top: 0.65em;
-  width: 12px;
-  height: 1px;
-  background: rgba(0, 0, 0, 0.45);
-}
-
-/* ── 2. Vision block — full width long-form ─────────────── */
-
-.vision {
-  /* Standalone slide content — no top margin/border (the previous
-     design had Pillars above; now Vision is the only block). */
-  margin-top: 0;
-  padding: 0;
-  border-top: none;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  flex: 1;
-  min-height: 0;
-}
-
-.subEyebrow {
-  font-family: var(--font-mono);
-  font-size: 11px;
-  letter-spacing: 0.26em;
-  text-transform: uppercase;
-  color: rgba(0, 0, 0, 0.55);
-}
-
-.subTitle {
-  font-family: var(--font-heading);
-  font-size: clamp(1.8rem, 1.4rem + 1.2vw, 2.6rem);
-  font-weight: 500;
-  letter-spacing: -0.022em;
-  line-height: 1.08;
-  color: #02000e;
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-  max-width: 24ch;
-}
-
-.subTitle em {
-  font-style: italic;
-  color: #02000e;
-}
-
-.visionLayout {
-  display: grid;
-  grid-template-columns: 1fr 1.05fr;
-  grid-template-rows: 1fr;
-  gap: 32px;
-  margin-top: 8px;
-  align-items: center;
-  min-height: 0;
-  flex: 1;
-}
-
-.visionCopy {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-.visionCopy p {
-  margin: 0;
-  font-size: 1rem;
-  line-height: 1.65;
-  color: rgba(0, 0, 0, 0.78);
-  max-width: 60ch;
-}
-
-.visionVisual {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 0;
-}
-
-/* Plate visual override — let the iso stack fill the column. */
-.visionPlate {
-  width: 100%;
-  height: 100%;
-  border: 1px solid rgba(0, 0, 0, 0.18);
-  background: transparent !important;
-}
-
-.visionPlateBody {
-  aspect-ratio: auto !important;
-  flex: 1 1 auto !important;
-  min-height: 0 !important;
-  max-height: calc(100vh - var(--nav-h, 64px) - 18vh);
-  padding: 0 !important;
-  background: transparent !important;
-}
-
-.visionPlateBody svg {
-  width: 100% !important;
-  height: 100% !important;
-  max-width: none !important;
-  max-height: none !important;
-  display: block;
-}
-
-@media (max-width: 720px) {
-  .visionLayout {
-    grid-template-columns: 1fr;
-    gap: 16px;
-  }
-}
-
-/* ── 3. Open & built-in-public — 3-cell grid ──────────────── */
-
-.open {
-  margin-top: 88px;
-  padding: 48px 0 0;
-  border-top: 1px solid rgba(0, 0, 0, 0.18);
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-.openGrid {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 0;
-  margin-top: 32px;
-  border-top: 1px solid rgba(0, 0, 0, 0.18);
-  border-left: 1px solid rgba(0, 0, 0, 0.18);
-}
-
-.openCell {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  padding: 32px 28px 36px;
-  border-right: 1px solid rgba(0, 0, 0, 0.18);
-  border-bottom: 1px solid rgba(0, 0, 0, 0.18);
-  min-width: 0;
-  background: var(--color-accent);
-  transition: background 0.25s var(--ease-out);
-}
-
-.openCell:hover {
-  background: rgba(0, 0, 0, 0.04);
-}
-
-.openLabel {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  color: rgba(0, 0, 0, 0.55);
-}
-
-.openText {
-  font-size: 0.95rem;
-  line-height: 1.55;
-  color: rgba(0, 0, 0, 0.78);
-  margin: 0;
-  flex: 1;
-}
-
-.openCta {
-  align-self: flex-start;
-  margin-top: 4px;
-}
-
-.openTeasers {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  border-top: 1px solid rgba(0, 0, 0, 0.12);
-  padding-top: 14px;
-}
-
-.openTeaserLink {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  text-decoration: none;
-  color: inherit;
-  padding: 8px 10px;
-  margin: 0 -10px;
-  transition: background 0.2s;
-}
-
-.openTeaserLink:hover {
-  background: rgba(0, 0, 0, 0.05);
-}
-
-.openTeaserDate {
-  font-family: var(--font-mono);
-  font-size: 9px;
-  letter-spacing: 0.2em;
-  color: rgba(0, 0, 0, 0.55);
-}
-
-.openTeaserTitle {
-  font-size: 0.86rem;
-  line-height: 1.35;
-  color: #02000e;
-  font-weight: 500;
-}
-
-/* ── Responsive ───────────────────────────────────────── */
-
-@media (max-width: 1100px) {
-  .pillars {
-    grid-template-columns: 1fr;
-  }
-  .openGrid {
-    grid-template-columns: 1fr;
-  }
-}
-
-@media (max-width: 600px) {
-  .pillar,
-  .openCell {
-    padding: 32px 24px;
-  }
-  .vision,
-  .open {
-    padding-top: 36px;
-    margin-top: 64px;
-  }
-}
