--- apps/landing/src/components/Section.module.css
@@ -1,139 +1,0 @@
-/**
- * Section — generic section frame from the design v2.
- * - Vertical rhythm 140px
- * - Border-top entre sections (sauf premier)
- * - .section-meta absolu en haut (S.0X · LABEL gauche, meta droite)
- * - variant peach flippe le bg + ink contrast
- */
-
-.section {
-  position: relative;
-  padding: 140px 0;
-  overflow: hidden;
-  background: var(--color-bg);
-}
-
-.section + .section,
-.section.firstAfterHero {
-  border-top: 1px solid var(--color-border-soft);
-}
-
-/* Peach variant — ink-on-peach. */
-.peach {
-  background: var(--color-accent);
-  color: #02000e;
-  border-top-color: rgba(0, 0, 0, 0.1);
-}
-
-/* Coordinate strip — absolute top, mono uppercase. */
-.meta {
-  position: absolute;
-  top: 32px;
-  left: 0;
-  right: 0;
-  display: flex;
-  justify-content: space-between;
-  padding: 0 var(--gutter);
-  font-family: var(--font-mono);
-  font-size: 10px;
-  font-weight: 500;
-  letter-spacing: 0.22em;
-  text-transform: uppercase;
-  color: var(--color-text-dim);
-  z-index: 2;
-  pointer-events: none;
-}
-
-.peach .meta {
-  color: rgba(0, 0, 0, 0.55);
-}
-
-.metaL,
-.metaR {
-  display: inline-flex;
-  align-items: center;
-  gap: 14px;
-}
-
-.dot {
-  width: 5px;
-  height: 5px;
-  background: var(--color-accent);
-  display: inline-block;
-}
-
-.peach .dot {
-  background: #02000e;
-}
-
-.body {
-  position: relative;
-  z-index: 2;
-}
-
-/* ── Blueprint draw-in: corner brackets + animated rule ──────── */
-
-.rule {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 1px;
-  background: var(--color-border-soft);
-  transform-origin: left center;
-  z-index: 3;
-  pointer-events: none;
-}
-
-.peach .rule {
-  background: rgba(0, 0, 0, 0.18);
-}
-
-/* The .section + .section { border-top } now lives in the animated rule
-   instead, so the rule can scaleX from 0. Remove the static border on
-   sections that have the rule. */
-.section + .section {
-  border-top: 0;
-}
-
-.corners {
-  position: absolute;
-  inset: 0;
-  pointer-events: none;
-  z-index: 3;
-}
-
-.corner {
-  position: absolute;
-  width: 22px;
-  height: 22px;
-}
-
-.corner_tl {
-  top: 16px;
-  left: 16px;
-}
-.corner_tr {
-  top: 16px;
-  right: 16px;
-}
-.corner_bl {
-  bottom: 16px;
-  left: 16px;
-}
-.corner_br {
-  bottom: 16px;
-  right: 16px;
-}
-
-@media (max-width: 768px) {
-  .section {
-    padding: 96px 0;
-  }
-  .meta {
-    top: 20px;
-    padding: 0 var(--gutter);
-    font-size: 9px;
-    letter-spacing: 0.18em;
-  }
-}
