--- apps/landing/src/components/SectionHead.module.css
@@ -1,64 +1,0 @@
-/**
- * SectionHead — eyebrow + title (+ sub) above each section's main content.
- * Two layouts:
- *   solo       — single column, lede stacked under title (max-width 720px)
- *   bipartite  — 2-col grid (1fr/1fr), title left, lede or right slot right
- */
-
-.head {
-  display: grid;
-  gap: 24px;
-  margin-bottom: 80px;
-}
-
-.solo {
-  grid-template-columns: 1fr;
-  max-width: 720px;
-}
-
-.bipartite {
-  grid-template-columns: minmax(0, 1fr) minmax(0, 1fr);
-  gap: 32px 64px;
-  align-items: end;
-}
-
-.left {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  min-width: 0;
-}
-
-.title {
-  margin: 0;
-}
-
-.peach .title,
-.peach .title :global(em) {
-  color: #02000e;
-}
-
-.right {
-  align-self: end;
-  padding-bottom: 8px;
-  display: flex;
-  justify-content: flex-end;
-}
-
-.right :global(.lede) {
-  margin-left: auto;
-}
-
-@media (max-width: 880px) {
-  .bipartite {
-    grid-template-columns: 1fr;
-    gap: 24px;
-    margin-bottom: 56px;
-  }
-  .right {
-    justify-content: flex-start;
-  }
-  .right :global(.lede) {
-    margin-left: 0;
-  }
-}
