--- apps/landing/src/components/SharedPlateA.module.css
@@ -1,50 +1,0 @@
-/**
- * SharedPlateA — Plate A element rendered once by the HexDeck and
- * animated across slide boundaries. Visual mirrors what used to be
- * Hero's `.diagram` + corner `.plateTag` spans.
- *
- * The HexDeck positions and animates this element via GSAP — this
- * stylesheet only handles the intrinsic look (aspect, tag layout).
- */
-
-.plate {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  aspect-ratio: 1;
-  pointer-events: none;
-  will-change: transform, opacity;
-}
-
-.plate > svg {
-  width: 100%;
-  height: 100%;
-  display: block;
-}
-
-.tag {
-  position: absolute;
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  text-transform: uppercase;
-  color: rgba(0, 0, 0, 0.55);
-  pointer-events: none;
-}
-
-.tl {
-  top: 0;
-  left: 0;
-}
-.tr {
-  top: 0;
-  right: 0;
-}
-.bl {
-  bottom: 0;
-  left: 0;
-}
-.br {
-  bottom: 0;
-  right: 0;
-}
