--- apps/landing/src/components/Steps.module.css
@@ -1,107 +1,0 @@
-/**
- * Steps (S.02) — single-col header, full-bleed PLATE.C (circles),
- * then 3-col grid of steps (1px shared borders).
- */
-
-.stepsLayout {
-  margin-bottom: 48px;
-  width: 100%;
-  position: relative;
-  z-index: 1;
-}
-
-.stepsPlate {
-  width: 100%;
-}
-
-/* Force the circles plate to widescreen aspect within Steps. */
-.stepsPlateBody {
-  aspect-ratio: 2 / 1 !important;
-  height: auto;
-}
-
-/* ── Steps grid ───────────────────────────────────────── */
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 0;
-  border-top: 1px solid var(--color-border-soft);
-  counter-reset: step;
-  position: relative;
-  z-index: 1;
-}
-
-.step {
-  position: relative;
-  padding: 48px 36px 56px;
-  border-right: 1px solid var(--color-border-soft);
-  border-bottom: 1px solid var(--color-border-soft);
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  min-height: 340px;
-  transition: background 0.3s var(--ease-out);
-}
-
-.step:last-child {
-  border-right: 0;
-}
-
-.step:hover {
-  background: var(--color-bg-alt);
-}
-
-.num {
-  font-family: var(--font-heading);
-  font-size: 3.6rem;
-  font-weight: 300;
-  line-height: 1;
-  color: var(--color-text-dim);
-  font-variation-settings: 'opsz' 144;
-  transition: color 0.25s;
-}
-
-.step:hover .num {
-  color: var(--color-accent-ink);
-}
-
-.name {
-  font-family: var(--font-heading);
-  font-size: 1.5rem;
-  font-weight: 500;
-  letter-spacing: -0.022em;
-  line-height: 1.15;
-  color: var(--color-text);
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-}
-
-.desc {
-  font-size: 0.95rem;
-  line-height: 1.6;
-  color: var(--color-text-muted);
-  flex: 1;
-  margin: 0;
-}
-
-.link {
-  align-self: flex-start;
-}
-
-.link.link:hover {
-  background: var(--color-accent) !important;
-  border-color: var(--color-accent) !important;
-  color: #02000e !important;
-}
-
-@media (max-width: 880px) {
-  .grid {
-    grid-template-columns: 1fr;
-  }
-  .step {
-    border-right: 0;
-  }
-}
