--- apps/landing/src/components/Values.module.css
@@ -1,145 +1,0 @@
-/**
- * Values (S.05) — 5-col ink-on-peach grid. Cards split by 1px ink
- * hairlines. Mono "VALUE NN", inline TRUST counter, Fraunces name,
- * body copy, and a pill outline vote button that fills ink on hover.
- */
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  grid-template-rows: 1fr;
-  gap: 0;
-  border-top: 1px solid rgba(255, 255, 255, 0.12);
-  border-left: 1px solid rgba(255, 255, 255, 0.12);
-  background: #02000e;
-  /* Take the slide's leftover height instead of growing with content. */
-  flex: 1;
-  min-height: 0;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 22px 18px 22px;
-  border-right: 1px solid rgba(255, 255, 255, 0.12);
-  border-bottom: 1px solid rgba(255, 255, 255, 0.12);
-  background: #02000e;
-  color: rgba(255, 255, 255, 0.92);
-  transition: background 0.25s;
-  min-height: 0;
-  overflow: hidden;
-}
-
-.card:hover {
-  background: rgba(255, 255, 255, 0.04);
-}
-
-.cardHead {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.num {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  color: rgba(255, 255, 255, 0.5);
-}
-
-.trust {
-  display: inline-flex;
-  align-items: baseline;
-  gap: 5px;
-  font-family: var(--font-mono);
-  color: var(--color-accent);
-  font-variant-numeric: tabular-nums;
-}
-
-.trustNum {
-  font-size: 13px;
-  font-weight: 700;
-}
-
-.trustUnit {
-  font-size: 9px;
-  letter-spacing: 0.16em;
-  text-transform: uppercase;
-  opacity: 0.7;
-}
-
-.name {
-  font-family: var(--font-heading);
-  font-size: 1.15rem;
-  font-weight: 500;
-  letter-spacing: -0.022em;
-  line-height: 1.15;
-  color: #ffffff;
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-}
-
-.desc {
-  font-size: 0.875rem;
-  line-height: 1.55;
-  color: rgba(255, 255, 255, 0.68);
-  flex: 1;
-  margin: 0;
-}
-
-.vote {
-  align-self: flex-start;
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  padding: 8px 14px;
-  border-radius: 999px;
-  border: 1px solid rgba(255, 255, 255, 0.7);
-  background: transparent;
-  color: rgba(255, 255, 255, 0.92);
-  font-family: var(--font-mono);
-  font-size: 10px;
-  font-weight: 500;
-  letter-spacing: 0.14em;
-  text-transform: uppercase;
-  cursor: pointer;
-  transition:
-    background 0.2s,
-    color 0.2s,
-    border-color 0.2s;
-}
-
-.vote:hover {
-  background: var(--color-accent);
-  color: #02000e;
-  border-color: var(--color-accent);
-}
-
-.vote:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-.error {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.18em;
-  text-transform: uppercase;
-  color: #c0392b;
-  margin: 4px 0 0;
-}
-
-@media (max-width: 720px) {
-  .grid {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media (max-width: 560px) {
-  .grid {
-    grid-template-columns: 1fr;
-  }
-}
