--- apps/landing/src/components/Videos.module.css
@@ -1,106 +1,0 @@
-/**
- * Videos (S.0B) — 2-column video grid. Asymmetric on desktop:
- * left video (long-form vision) takes more width than right (short
- * walkthrough). Stacks on tablet + mobile.
- */
-
-.grid {
-  display: grid;
-  grid-template-columns: 1.4fr 1fr;
-  gap: 32px;
-  margin-top: 56px;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  background: transparent;
-}
-
-.frame {
-  position: relative;
-  aspect-ratio: 16 / 9;
-  border: 1px solid var(--color-border-soft);
-  background: rgba(255, 255, 255, 0.02);
-  overflow: hidden;
-}
-
-.video {
-  width: 100%;
-  height: 100%;
-  display: block;
-  object-fit: cover;
-}
-
-.placeholder {
-  position: absolute;
-  inset: 0;
-  display: grid;
-  place-items: center;
-  gap: 8px;
-  color: var(--color-text-dim);
-  background:
-    repeating-linear-gradient(
-      45deg,
-      transparent 0 18px,
-      rgba(255, 255, 255, 0.025) 18px 19px
-    );
-}
-
-.placeholderTag {
-  font-family: var(--font-mono);
-  font-size: 14px;
-  letter-spacing: 0.28em;
-  color: var(--color-text);
-}
-
-.placeholderHint {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.18em;
-  text-transform: uppercase;
-  opacity: 0.6;
-}
-
-.caption {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 4px 0 0;
-}
-
-.tag {
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  color: var(--color-text-dim);
-  text-transform: uppercase;
-}
-
-.title {
-  font-family: var(--font-heading);
-  font-size: 1.4rem;
-  font-weight: 500;
-  letter-spacing: -0.022em;
-  line-height: 1.15;
-  color: var(--color-text);
-  font-variation-settings:
-    'opsz' 144,
-    'SOFT' 30;
-  margin: 0;
-}
-
-.desc {
-  font-size: 0.95rem;
-  line-height: 1.55;
-  color: var(--color-text-muted);
-  max-width: 56ch;
-  margin: 0;
-}
-
-@media (max-width: 1000px) {
-  .grid {
-    grid-template-columns: 1fr;
-  }
-}
