--- apps/landing/src/proto/SharedPlate.module.css
@@ -1,46 +1,0 @@
-/**
- * SharedPlate — internal layout (the wrapper is positioned by
- * `Deck.module.css .sharedPlate`).
- */
-
-.plate {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  aspect-ratio: 1;
-  pointer-events: none;
-  will-change: transform, opacity;
-}
-
-.plate > svg {
-  width: 100%;
-  height: 100%;
-  display: block;
-}
-
-.tag {
-  position: absolute;
-  font-family: var(--font-mono);
-  font-size: 10px;
-  letter-spacing: 0.22em;
-  text-transform: uppercase;
-  color: rgba(0, 0, 0, 0.55);
-  pointer-events: none;
-}
-
-.tl {
-  top: 0;
-  left: 0;
-}
-.tr {
-  top: 0;
-  right: 0;
-}
-.bl {
-  bottom: 0;
-  left: 0;
-}
-.br {
-  bottom: 0;
-  right: 0;
-}
